borgmatic: add option for pattern matching
Borgmatic has support for Borg's pattern matching in backups.
It is mutually exclusive with the existing `sourceDirectories` option,
but pattern matching is more granular, so as a reasonable default
`programs.borgmatic.backups.<name>.location.sourceDirectories` is set to
null if both it and `programs.borgmatic.backups.<name>.location.patterns`
are specified.